refactor(frontend): extract shared NavTabs, tanstack table setup, and userscript install step; reuse Tooltip in FormLabelWithTooltip
